feat(hubble): nats - decouple block processing - rework: move util
